use new repository for work items
