+++ backend/event-listeners/src/infrastructure/database/mod.rs
@@ -7,10 +7,8 @@ mod indexer;
 mod project;
 mod project_github_repos;
 mod project_lead;
-mod work_item;
 
 pub use github_issues::Repository as GithubIssuesRepository;
 pub use project::Repository as ProjectRepository;
 pub use project_github_repos::Repository as ProjectGithubReposRepository;
 pub use project_lead::Repository as ProjectLeadRepository;
-pub use work_item::Repository as WorkItemRepository;
